added setDirection function to controlls
I needed setDirection and found it in the 'controls' Types (AnimationDispatch) but it seems it was never implemented. I simply followed the existing code pattern and added it in. After a quick glance I could not get a way to show current play direction. It is in the controls object.