Merging setDirection enhancements into Master (#12)
* added setDirection function to controlls

I needed setDirection and found it in the 'controls' Types (AnimationDispatch) but it seems it was never implemented. I simply followed the existing code pattern and added it in. After a quick glance I could not get a way to show current play direction. It is in the controls object.

* Added setDirection to dependency array

* enhancements: setDirection.
moved npm package readme into proper directory.
added abstraction UseLottieState.
Added playDirection to state.

* bumping up to v0.1.1.
Added contributor - Arno (H) Welgemoed.

Co-authored-by: ahwelgemoed <arno.welgemoed@outlook.com>
Co-authored-by: Noam Gabriel Jacobson <njacobson@developertown.com>